// src/app/model/event.ts
import {Vendor} from './vendor';

export class Event {
  id: number;
  eName: string;
  eDate: string;
  eCname: string;
  eDesc: string;
  ePlace: string;
  eStatus: string;
  eAny: string;
  eMeet: boolean;
  eServices: string;


  // private long id;
  // private String eName;
  // private String eDesc;
  // private String eCname;
  // private String eAny;
  // private String ePlace;
  // private String eServices;
  // private Date eDate;
  // private boolean eMeet;
  // private String eStatus;
}